ConsoleZaberCommander
Base Code for future integration -- This communicates with the Zaber actuators and avoids are known "deadzones" or, where we don't want the actuators to go because our tweezers might run into something.  If the default of moving things x, then y, then z positioning will run the tweezers into something, it finds a different path.  CANNOT move to the other side of an obstacle - since I probably won't need to be doing this with my setup, I haven't programmed a way to handle this case.