Joystick + Live-chart
Fixed the joysticks so just outside deadzone is smooth.  Data logger now writes to file Arduino sensor input.  Live-chart was added as well for (almost!) real-time feedback from the Arduino sensors -- it is lagging by about 3 seconds as a result of the window loading and hitting the inject data button.